requirements.txt
# Core Python dependencies
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# OpenAI API
openai==0.28.1

# Google Sheets integration
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Utilities
requests==2.31.0
